requirements.txt
googlegenerativeai
openai
pandas
requests
ast
io
unittest
re